requirements.txt
aiohttp==3.9.3
aiohttp-retry==2.8.3
aiosignal==1.3.1
attrs==23.2.0
certifi==2024.2.2
charset-normalizer==3.3.2
frozenlist==1.4.1
idna==3.7
iniconfig==2.0.0
multidict==6.0.5
packaging==24.0
pluggy==1.4.0
PyJWT==2.8.0
pytest==8.1.1
python-dotenv==1.0.1
requests==2.31.0
twilio==9.0.4
urllib3==2.2.1
yarl==1.9.4


# pip3 install twilio
#pip install python-dotenv